microsalt refactor wip
